Added IndexedSelector
This will simplify many future selectors.
