+++ src/main/java/dte/desktobeauty/wallpaper/selector/GradualRandomSelector.java
@@ -2,9 +2,7 @@ package dte.desktobeauty.wallpaper.selector;
 
 import dte.desktobeauty.wallpaper.Wallpaper;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.concurrent.ThreadLocalRandom;
 import java.util.stream.IntStream;
 
@@ -12,10 +10,8 @@ import java.util.stream.IntStream;
  * Selects a random wallpaper whose brightness is moderately higher than the previous one;
  * This prevents the annoying flash when a bright picture is selected right after a dark one.
  */
-public class GradualRandomSelector extends AbstractSelector
+public class GradualRandomSelector extends IndexedSelector
 {
-    private final Map<List<Wallpaper>, Integer> currentIndexes = new HashMap<>();
-
     private static final int MAX_BRIGHTNESS_DIFFERENCE = 20;
 
     public GradualRandomSelector()
@@ -24,38 +20,21 @@ public class GradualRandomSelector extends AbstractSelector
     }
 
     @Override
-    public Wallpaper selectFrom(List<Wallpaper> wallpapers)
+    protected int selectFirstIndex(List<Wallpaper> wallpapers)
     {
-        int nextIndex = calculateNextIndex(wallpapers);
-        this.currentIndexes.put(wallpapers, nextIndex);
-
-        return wallpapers.get(nextIndex);
+        return ThreadLocalRandom.current().nextInt(wallpapers.size());
     }
 
-    private int calculateNextIndex(List<Wallpaper> wallpapers)
+    @Override
+    protected int selectNextIndex(List<Wallpaper> wallpapers, int currentIndex)
     {
-        Integer currentIndex = this.currentIndexes.get(wallpapers);
-
-        //first selection is random
-        if(currentIndex == null)
-            return ThreadLocalRandom.current().nextInt(wallpapers.size());
-
         Wallpaper currentWallpaper = wallpapers.get(currentIndex);
 
-        //choose a random candidate and return its index
         int[] candidateIndexes = IntStream.range(0, wallpapers.size())
-                .filter(i -> isCandidate(currentWallpaper, wallpapers.get(i)))
+                .filter(i -> i != currentIndex) //skip the current wallpaper
+                .filter(i -> currentWallpaper.getBrightnessDifferenceTo(wallpapers.get(i)) <= MAX_BRIGHTNESS_DIFFERENCE)
                 .toArray();
 
         return candidateIndexes[ThreadLocalRandom.current().nextInt(candidateIndexes.length)];
-    }
-
-    private static boolean isCandidate(Wallpaper currentWallpaper, Wallpaper wallpaper)
-    {
-        //skip the current wallpaper
-        if(currentWallpaper.equals(wallpaper))
-            return false;
-
-        return currentWallpaper.getBrightnessDifferenceTo(wallpaper) <= MAX_BRIGHTNESS_DIFFERENCE;
     }
 }
